split filter widget into multiple files
\to make it more manageable
