+++ src/admc/CMakeLists.txt
@@ -42,6 +42,8 @@ add_executable(admc
     ad_object.cpp
 
     filter_widget/filter_widget.cpp
+    filter_widget/filter_widget_normal_tab.cpp
+    filter_widget/filter_widget_advanced_tab.cpp
     
     tabs/details_tab.cpp
     tabs/general_tab.cpp
